requirements.txt
yfinance==0.2.51
requests==2.32.3
PyYAML==6.0.2
pytz==2024.2
python-telegram-bot==21.6
